requirements.txt
pillow
praw
